+++ Server/include/SphereManager.hpp
@@ -43,30 +43,30 @@ namespace SphereSim{
 		 * \param s Sphere to add.
 		 * \return Current sphere count.
 		 */
-		unsigned int addSphere(Sphere s);
+		quint16 addSphere(Sphere s);
 		/**
 		 * \brief Adds a new sphere.
 		 * \return Current sphere count.
 		 */
-		unsigned int addSphere();
+		quint16 addSphere();
 		
 		/**
 		 * \brief Removes the last sphere.
 		 * \return Current sphere count.
 		 */
-		unsigned int removeLastSphere();
+		quint16 removeLastSphere();
 		/**
 		 * \brief Removes a specific sphere.
 		 * \param i Index of the sphere to remove.
 		 * \return Current sphere count.
 		 */
-		unsigned int removeSphere(unsigned int i);
+		quint16 removeSphere(quint16 i);
 		
 		/**
 		 * \brief Returns the current sphere count.
 		 * \return Current sphere count.
 		 */
-		unsigned int getCount();
+		quint16 getCount();
 		
 		/**
 		 * \brief Updates a specific sphere.
@@ -74,20 +74,20 @@ namespace SphereSim{
 		 * \param s Sphere data to update.
 		 * \return Current sphere count.
 		 */
-		unsigned int updateSphere(unsigned int i, Sphere s);
+		quint16 updateSphere(quint16 i, Sphere s);
 		
 		/**
 		 * \brief Gets a specific sphere.
 		 * \param i Index of the sphere to get.
 		 * \return Copy of the requested sphere.
 		 */
-		Sphere getSphere(unsigned int i);
+		Sphere getSphere(quint16 i);
 		
 		/**
 		 * \brief Calculates the sphere movements for one step.
 		 * \return Millisecond time used to calculate the step.
 		 */
-		unsigned int calculateStep();
+		quint16 calculateStep();
 		
 		friend class ActionReceiver;
 	};
